[#713] app boilerplate to files (#862)
* create app/modules.go

* add AppKeepers struct
